modify cart step
